requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
djangorestframework==3.12.4
gunicorn==20.1.0
html5lib==1.1
idna==3.2
lxml==4.6.3
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==8.3.1
psycopg2-binary==2.9.1
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
Wand==0.6.6
webencodings==0.5.1
